Major Logic changes:
pf.c: Adding an option to turn the NESTED_DUAL on/off

amcl_laser.cpp:

-NestedBeamModel:
--Removed the mapping of any distance greater than max_occ_dist to
max_occ_dist...instead it extends as long as it can.

-LikelihoodFieldModel_AW:
--Only using the beams which have nested particles within max_occ_dist
of the hit-point for advanced weighting. otherwise they default to
normal weighting now.
--Using the nearest nested particle within max_occ_dist of the hit-point
of the beam for weighting instead of the old kernel function.