Changes selectively copied from SNPF-AW, with modifications:
-Copied LikelihoodFieldModel_AW...it now selects only nested particles
within max_occ_dist to use in Advanced Weighting. Also, we select the
closest nested particle instead of using the kernel function. Also,
optimized this process a little to make it faster, and allow ANPF-AW to
run with 5000 particles.
Changed pf.c to allow NESTED_DUAL deciding whether or not to allow
mixture MCL in nested particles. Also, optimized this using #if to
reduce no of conditional checks to make ANPF-AW faster, so as to run
with 5000 particles as well.